docs: move sorting code to a new module
